+++ src/background/modules/notifications/application.rs
@@ -28,9 +28,8 @@ use crate::{
     event_manager, log_error,
     modules::uwp::get_hightest_quality_posible,
     seelen::get_app_handle,
-    seelen_weg::icon_extractor::extract_and_save_icon_umid,
     trace_lock,
-    utils::{convert_file_to_src, spawn_named_thread},
+    utils::{convert_file_to_src, icon_extractor::extract_and_save_icon_umid, spawn_named_thread},
     windows_api::traits::EventRegistrationTokenExt,
 };
 
@@ -215,7 +214,7 @@ impl NotificationManager {
                 "http" | "https" => {}
                 "ms-appx" | "ms-appx-web" => {
                     let path = package_path.clone()?.join(uri_path);
-                    if let Some(path) = get_hightest_quality_posible(&path) {
+                    if let Some((path, _)) = get_hightest_quality_posible(&path) {
                         log::debug!("  Resolved path: {}", path.display());
                         image.src = convert_file_to_src(&path);
                     } else {